# Conflicts:
#	app/build.gradle
#	app/src/main/java/com/patsurvey/nudge/activities/settings/SettingScreen.kt
#	app/src/main/java/com/patsurvey/nudge/activities/settings/domain/use_case/ClearLocalDBUseCase.kt
#	app/src/main/java/com/patsurvey/nudge/activities/settings/domain/use_case/GetUserDetailsUseCase.kt
#	app/src/main/java/com/patsurvey/nudge/activities/settings/domain/use_case/LogoutUseCase.kt
#	app/src/main/java/com/patsurvey/nudge/activities/ui/progress/VillageSelectionViewModel.kt
#	app/src/main/java/com/patsurvey/nudge/navigation/baseline/BSHomeNavScreen.kt
#	app/src/main/java/com/patsurvey/nudge/navigation/baseline/BSNavHomeGraph.kt
#	app/src/main/java/com/patsurvey/nudge/navigation/selection/NavHomeGraph.kt
#	baselinesurvey/build.gradle
#	baselinesurvey/src/main/AndroidManifest.xml
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/AppConstants.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/data/prefs/PrefRepo.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/data/prefs/SharedPrefs.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/di/BaselineModule.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/di/NetworkModule.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/Constants/QuestionType.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/auth/presentation/OtpVerificationScreenComponent.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/common_components/common_setting/CommonSettingScreen.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/form_response_summary_screen/presentation/FormQuestionSummaryScreen.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/mission_screen/presentation/MissionScreen_1.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/mission_screen/viewmodel/MissionViewModel.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/mission_summary_screen/presentation/MissionSummaryScreen.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/question_screen/presentation/QuestionScreen.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/question_screen/presentation/questionComponent/NestedLazyList.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/search/presentation/SearchScreen.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/section_screen/presentation/SectionListScreen.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/setting/domain/repository/SettingBSRepository.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/setting/domain/repository/SettingBSRepositoryImpl.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/setting/domain/use_case/SettingBSUserCase.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/setting/presentation/SettingBSScreen.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/setting/viewmodel/SettingBSViewModel.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/splash/domain/repository/SplashScreenRepositoryImpl.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/start_screen/viewmodel/BaseLineStartViewModel.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/surveyee_screen/domain/repository/DataLoadingScreenRepository.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/surveyee_screen/domain/repository/DataLoadingScreenRepositoryImpl.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/surveyee_screen/domain/use_case/FetchSurveyFromNetworkUseCase.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/surveyee_screen/presentation/DataLoadingScreenComponent.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/surveyee_screen/presentation/SurveyeeListScreen.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/surveyee_screen/viewmodel/DataLoadingScreenViewModel.kt
#	baselinesurvey/src/main/java/com/nrlm/baselinesurvey/utils/BaselineCore.kt
#	core/src/main/java/com/nudge/core/CoreConstants.kt

+++ baselinesurvey/src/main/java/com/nrlm/baselinesurvey/ui/mission_screen/domain/use_case/FetchMissionDataFromNetworkUseCase.kt
@@ -4,10 +4,15 @@ import com.nrlm.baselinesurvey.SUCCESS
 import com.nrlm.baselinesurvey.database.entity.ActivityTaskEntity
 import com.nrlm.baselinesurvey.database.entity.MissionActivityEntity
 import com.nrlm.baselinesurvey.database.entity.MissionEntity
+import com.nrlm.baselinesurvey.network.ApiException
+import com.nrlm.baselinesurvey.network.SUBPATH_GET_MISSION
 import com.nrlm.baselinesurvey.ui.surveyee_screen.domain.repository.DataLoadingScreenRepository
 import com.nrlm.baselinesurvey.utils.BaselineLogger
+import com.nudge.core.enums.ApiStatus
 import kotlinx.coroutines.delay
-
+import com.nrlm.baselinesurvey.BLANK_STRING
+import com.nrlm.baselinesurvey.DEFAULT_ERROR_CODE
+import com.nrlm.baselinesurvey.DEFAULT_SUCCESS_CODE
 class FetchMissionDataFromNetworkUseCase(
     private val repository: DataLoadingScreenRepository
 ) {
@@ -17,17 +22,29 @@ class FetchMissionDataFromNetworkUseCase(
     }
     suspend operator fun invoke(): Boolean {
         try {
+            if (!repository.isNeedToCallApi(SUBPATH_GET_MISSION)) {
+                return false
+            }
+            repository.insertApiStatus(SUBPATH_GET_MISSION)
+
             val apiResponse = repository.fetchMissionDataFromServer("en", "BASELINE")
             if (apiResponse.status.equals(SUCCESS, true)) {
                 apiResponse.data?.let { missionApiResponse ->
-                    repository.deleteMissionsFromDB()
+                    repository.updateApiStatus(
+                        SUBPATH_GET_MISSION,
+                        status = ApiStatus.SUCCESS.ordinal,
+                        BLANK_STRING,
+                        DEFAULT_SUCCESS_CODE
+                    )
+                    /*repository.deleteMissionsFromDB()
                     repository.deleteMissionActivitiesFromDB()
-                    repository.deleteActivityTasksFromDB()
+                    repository.deleteActivityTasksFromDB()*/
                     missionApiResponse.forEach { mission ->
                         var activityTaskSize = 0
                         mission.activities.forEach { activity ->
                             repository.saveMissionsActivityToDB(
                                 MissionActivityEntity.getMissionActivityEntity(
+                                    userId = repository.getBaseLineUserId(),
                                     missionId = mission.missionId,
                                     activity = activity,
                                     activityTaskSize = activity.tasks.size
@@ -37,6 +54,7 @@ class FetchMissionDataFromNetworkUseCase(
                                 if (task.id != null) {
                                     repository.saveActivityTaskToDB(
                                         ActivityTaskEntity.getActivityTaskEntity(
+                                            userId = repository.getBaseLineUserId(),
                                             missionId = mission.missionId,
                                             activityId = activity.activityId,
                                             activityName = activity.activityName,
@@ -50,6 +68,7 @@ class FetchMissionDataFromNetworkUseCase(
                         delay(100)
                         repository.saveMissionToDB(
                             MissionEntity.getMissionEntity(
+                                userId = repository.getBaseLineUserId(),
                                 activityTaskSize = activityTaskSize,
                                 mission = mission
                             )
@@ -59,11 +78,31 @@ class FetchMissionDataFromNetworkUseCase(
                 }
                 return false
             } else {
+                repository.updateApiStatus(
+                    SUBPATH_GET_MISSION,
+                    status = ApiStatus.FAILED.ordinal,
+                    apiResponse.message ?: BLANK_STRING,
+                    DEFAULT_ERROR_CODE
+                )
                 return false
             }
+        } catch (apiException: ApiException) {
+            repository.updateApiStatus(
+                SUBPATH_GET_MISSION,
+                status = ApiStatus.FAILED.ordinal,
+                apiException.message ?: BLANK_STRING,
+                apiException.getStatusCode()
+            )
+            throw apiException
         } catch (ex: Exception) {
-            BaselineLogger.e("FetchSurveyFromNetworkUseCase", "invoke", ex)
-            return false
+            repository.updateApiStatus(
+                SUBPATH_GET_MISSION,
+                status = ApiStatus.FAILED.ordinal,
+                ex.message ?: BLANK_STRING,
+                DEFAULT_ERROR_CODE
+            )
+            BaselineLogger.e("FetchUserDetailFromNetworkUseCase", "invoke", ex)
+            throw ex
         }
     }
